factor out editHTMLs
